+++ src/test/scala/com/scalableminds/fossildb/FossilDBSuite.scala
@@ -69,6 +69,16 @@ class FossilDBSuite extends AnyFlatSpec with BeforeAndAfterEach with TestHelpers
     assert(testData2 == reply.value)
   }
 
+  "PutMultipleVersions" should "overwrite old values, leave others untouched" in {
+    client.put(PutRequest(collectionA, aKey, Some(0), testData1))
+    client.put(PutRequest(collectionA, aKey, Some(2), testData1))
+    client.putMultipleVersions(PutMultipleVersionsRequest(collectionA, aKey, Seq(1,2,3), Seq(testData2, testData3, testData3)))
+    val reply = client.getMultipleVersions(GetMultipleVersionsRequest(collectionA, aKey))
+    assert(reply.values.length == 4)
+    assert(reply.versions == Seq(3,2,1,0))
+    assert(reply.values == Seq(testData3, testData3, testData2, testData1))
+  }
+
   it should "fail on non-existent collection" in {
     val reply = client.put(PutRequest("nonExistentCollection", aKey, Some(0), testData1))
     assert(!reply.success)
@@ -132,6 +142,21 @@ class FossilDBSuite extends AnyFlatSpec with BeforeAndAfterEach with TestHelpers
     client.delete(DeleteRequest(collectionA, aKey, 1))
     val reply = client.get(GetRequest(collectionA, aKey, Some(1)))
     assert(testData1 == reply.value)
+  }
+
+  "DeleteAllByPrefix" should "delete all versions of all values matching this prefix" in {
+    client.put(PutRequest(collectionA, "prefixedA", Some(0), testData1))
+    client.put(PutRequest(collectionA, "prefixedA", Some(1), testData1))
+    client.put(PutRequest(collectionA, "prefixedB", Some(0), testData2))
+    client.put(PutRequest(collectionA, "prefixedC", Some(0), testData2))
+    client.put(PutRequest(collectionA, "differentKey", Some(0), testData2))
+    client.put(PutRequest(collectionA, "differentKey", Some(1), testData2))
+    client.put(PutRequest(collectionA, "yetDifferentKey", Some(0), testData2))
+    client.deleteAllByPrefix(DeleteAllByPrefixRequest(collectionA, "prefixed"))
+    val reply = client.listKeys(ListKeysRequest(collectionA))
+    assert(reply.keys.length == 2)
+    assert(reply.keys.contains("differentKey"))
+    assert(reply.keys.contains("yetDifferentKey"))
   }
 
   "ListKeys" should "list all keys of a collection" in {